service: flatten kvm controller
This PR basically gets rid of `v22` in `service/controller/kvm/v22`
package and its sub-packages.
