+++ service/controller/internal/configmap/configmap.go
@@ -13,7 +13,7 @@ import (
 	awskey "github.com/giantswarm/cluster-operator/service/controller/aws/v22/key"
 	azurekey "github.com/giantswarm/cluster-operator/service/controller/azure/v22/key"
 	"github.com/giantswarm/cluster-operator/service/controller/key"
-	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/v22/key"
+	kvmkey "github.com/giantswarm/cluster-operator/service/controller/kvm/key"
 )
 
 // Config represents the configuration used to create a new configmap service.
