+++ service/version_bundles.go
@@ -5,7 +5,7 @@ import (
 
 	awsv22 "github.com/giantswarm/cluster-operator/service/controller/aws/v22"
 	azurev22 "github.com/giantswarm/cluster-operator/service/controller/azure/v22"
-	kvmv22 "github.com/giantswarm/cluster-operator/service/controller/kvm/v22"
+	"github.com/giantswarm/cluster-operator/service/controller/kvm"
 )
 
 func NewVersionBundles(p string) []versionbundle.Bundle {
@@ -13,7 +13,7 @@ func NewVersionBundles(p string) []versionbundle.Bundle {
 
 	versionBundles = append(versionBundles, awsv22.VersionBundle())
 	versionBundles = append(versionBundles, azurev22.VersionBundle())
-	versionBundles = append(versionBundles, kvmv22.VersionBundle())
+	versionBundles = append(versionBundles, kvm.VersionBundle())
 
 	return versionBundles
 }
